requirements.txt

fastapi
uvicorn
python-multipart
python-dotenv
requests
pypandoc
mammoth
openai
# uvloop removed as it's not supported on Windows
# httptools removed as it's optional
